Local Web server: Should be optional

+++ spec/support/local_web_server.rb
@@ -9,21 +9,23 @@ module LocalWebServer
     end
 
     config.before :all, type: :feature do
-      puts "Bringing up local webserver on port #{port}"
-      FileUtils.cd server_root_path do
-        # NOTE: Somehow puts or $stdout or $stderr do not work
-        # when running in `Rake::Task['foo'].invoke` context
-        # puts "Running webserver in #{Dir.pwd}"
+      if port > 0
+        puts "Bringing up local webserver on port #{port}"
+        FileUtils.cd server_root_path do
+          # NOTE: Somehow puts or $stdout or $stderr do not work
+          # when running in `Rake::Task['foo'].invoke` context
+          # puts "Running webserver in #{Dir.pwd}"
 
-        $SERVER_PID = fork do
-          exec "ruby -run -e httpd . -p #{port}"
+          $SERVER_PID = fork do
+            exec "ruby -run -e httpd . -p #{port}"
+          end
         end
-      end
 
-      at_exit do
-        puts "Bringing down local webserver"
-        Process.kill 'TERM', $SERVER_PID
-        Process.wait $SERVER_PID
+        at_exit do
+          puts "Bringing down local webserver"
+          Process.kill 'TERM', $SERVER_PID
+          Process.wait $SERVER_PID
+        end
       end
     end
   end
